requirements.txt
streamlit
torch
llama-index
gradientai
python-dotenv
transformers
sentencepiece
pypdf